30/01/2018

+++ AionHR.Model/Employees/Profile/EmployeeQuickView.cs
@@ -27,10 +27,10 @@ namespace AionHR.Model.Employees.Profile
         public int indemnity { get; set; }
         public DateTime? lastLeaveStartDate { get; set; }
         public DateTime? lastLeaveEndDate { get; set; }
-        public short usedLeavesLeg { get; set; }
+        public double usedLeavesLeg { get; set; }
 
         public string serviceDuration { get; set; }
-        public short leaveBalance { get; set; }
+        public double leaveBalance { get; set; }
         public short earnedLeavesLeg { get; set; }
 
         public string LastLeave(string format)
@@ -53,7 +53,7 @@ namespace AionHR.Model.Employees.Profile
         public string countryName { get; set; }
         public DateTime? hireDate { get; set; }
         public int earnedLeaves { get;  set; }
-        public int usedLeaves { get; set;  }
+        public double usedLeaves { get; set;  }
         public int paidLeaves { get; set; }
         public int salary { get; set; }
         public string currencyName { get; set; }
